+++ docs/en/04_Changelogs/4.0.0.md
@@ -28,6 +28,7 @@
  * `File` is now versioned, and should be published before they can be used on the frontend.
    See section on [Migrating File DataObject from 3.x to 4.0](#migrating-file-dataobject-from-3x-to-40)
    below for upgrade notes.
+ * Removed `dev/tests/` controller in favour of standard `vendor/bin/phpunit` command
 
 ## New API
 
@@ -47,6 +48,11 @@
     * `Configurable` Provides Config API helper methods
     * `Injectable` Provides Injector API helper methods
     * `Extensible` Allows extensions to be applied
+ * Removed ability to run tests via web requests (`http://mydomain.com/dev/tests`), use the standard CLI command instead (`vendor/bin/phpunit`)
+ * Removed `dev/jstests/` controller (no replacement)
+ * Moved test database cleanup task from `sake dev/tests/cleanupdb` to `sake dev/tasks/CleanupTestDatabasesTask`
+ * Removed `TestRunner` and `JSTestRunner` APIs
+ * Removed `PhpUnitWrapper`, `PhpUnitWrapper_3_4`, `PhpUnitWrapper_3_5`, `PhpUnitWrapper_Generic`, `SapphireTestSuite` APIs
  * `HtmlEditorConfig` is now an abstract class, with a default implementation `TinyMCEConfig` for the built in
    TinyMCE editor.
  * `HtmlEditorField::setEditorConfig` may now take an instance of a `HtmlEditorConfig` class, as well as a
