+++ tests/bootstrap.php
@@ -7,12 +7,6 @@
 // Make sure display_errors is on
 ini_set('display_errors', 1);
 
-// Check we're using at least PHPUnit 3.5
-if(version_compare(PHPUnit_Runner_Version::id(), '3.5', '<')) {
-	echo 'PHPUnit 3.5 required to run tests using bootstrap.php';
-	die();
-}
-
 // Fake the script name and base
 global $_SERVER;
 if (!$_SERVER) $_SERVER = array();
@@ -50,7 +44,7 @@ global $databaseConfig;
 DB::connect($databaseConfig);
 
 // Now set a fake REQUEST_URI
-$_SERVER['REQUEST_URI'] = BASE_URL . '/dev';
+$_SERVER['REQUEST_URI'] = BASE_URL;
 
 // Fake a session
 $_SESSION = null;
@@ -58,8 +52,7 @@ $_SESSION = null;
 // Prepare manifest autoloader
 $controller = new FakeController();
 
-// Get test manifest
-TestRunner::use_test_manifest();
+SapphireTest::use_test_manifest();
 
 SapphireTest::set_is_running_test(true);
